pyproject.toml
[project]
name = "aether"
version = "0.1.0"
description = "An autonomous, self-evolving agent orchestration framework based on SCION."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform",
    "pydantic",
    "python-dotenv",
    "rich",
    "typer",
    "uvicorn",
    "fastapi",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.uv.workspace]
members = [
    "aether_tuning",
]
